update owner ship

+++ src/main/java/funix/epfw/controller/home/ProcessContact.java
@@ -1,7 +1,9 @@
 package funix.epfw.controller.home;
 
+import funix.epfw.constants.AuthUtil;
 import funix.epfw.model.Contact;
 import funix.epfw.service.home.ContactService;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -20,13 +22,22 @@ public class ProcessContact {
     }
 
     @GetMapping("/deleteContact/{id}")
-    public String deleteContact(@PathVariable Long id) {
+    public String deleteContact(@PathVariable Long id, HttpSession session) {
+        String checkAuth = AuthUtil.checkAdminAuth(session);
+        if (checkAuth != null) {
+            return checkAuth;
+        }
         contactService.deleteContact(id);
         return "redirect:/manageContact";
     }
 
     @GetMapping("/doneContact/{id}")
-    public String doneContact(@PathVariable Long id) {
+    public String doneContact(@PathVariable Long id, HttpSession session) {
+        String checkAuth = AuthUtil.checkAdminAuth(session);
+        if (checkAuth != null) {
+            return checkAuth;
+        }
+
         Contact contact = contactService.getContactById(id);
         contact.setUpdatedTime(LocalDateTime.now());
         contactService.doneContact(id);
